Adding pundit

+++ app/controllers/comments_controller.rb
@@ -25,16 +25,13 @@ class CommentsController < ApplicationController
   end
 
   def edit
-    @comment = current_user.comments.find_by(id: params[:id])
-
-    if @comment.nil?
-      flash[:warning] = "You don't have permissions to edit this comment."
-      redirect_to root_path
-    end
+    @comment = Comment.find_by_id params[:id]
+    authorize @comment
   end
 
   def update
-    @comment = current_user.comments.find_by(id: params[:id])
+    @comment =  Comment.find_by_id params[:id]
+    authorize @comment
     if @comment && @comment.update(comment_params)
       flash[:success] = "Comment updated successfully."
       redirect_to @comment.parent_id.present? ? comment_path(@comment.parent_id) : comments_path
@@ -44,12 +41,11 @@ class CommentsController < ApplicationController
   end
 
   def destroy
-    comment = current_user.comments.find_by(id: params[:id])
+    comment =  Comment.find_by_id params[:id]
+    authorize comment
     if comment
       comment.destroy
       flash[:success] = "Commet deleted successfully."
-    else
-      flash[:danger] = "You don't have permissions to delete this comment."
     end
     redirect_to comments_path
   end
